@@ -131,32 +130,32 @@
     "lodash": "^4.17.21",
     "nanostores": "^0.7.4",
     "query-string": "^9.0.0",
-    "react": "18.2.0",
+    "react": "18.3.1",
     "react-app-rewired": "^2.2.1",
     "react-confirm": "^0.2.3",
-    "react-dom": "18.2.0",
+    "react-dom": "18.3.1",
     "react-ga4": "^2.1.0",
     "react-icons": "^4.12.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-input-mask": "^3.0.0-alpha.2",
     "react-refresh": "^0.14.0",
-    "react-router-dom": "^6.22.3",
+    "react-router-dom": "6.11.2",
     "react-scripts": "^5.0.1",
     "react-table": "^7.8.0",
     "react-test-renderer": "18.2.0",
-    "solid-js": "1.6.11",
+    "solid-js": "1.9.3",
     "solid-toast": "^0.5.0",
-    "tailwindcss": "^3.3.2",
+    "tailwindcss": "3.4.3",
     "terracotta": "^1.0.3",
-    "ts-jest": "^29.0.5",
+    "ts-jest": "29.1.5",
     "ts-node": "10.9.1",
     "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.6.3",
     "ulid": "^2.3.0",
     "url-loader": "^4.1.1",
     "use-debounce": "^9.0.3",
     "web-vitals": "^3.3.0",
-    "webpack": "^5.90.3",
+    "webpack": "5.95.0",
     "yup": "^0.32.11",
     "zod": "^3.22.4"
   },
