@@ -11,8 +11,8 @@ import type {
   Treatment
 } from '@photonhealth/sdk/dist/types';
 import gql from 'graphql-tag';
-import { GraphQLError } from 'graphql';
 import jwtDecode from 'jwt-decode';
+import type { GraphQLFormattedError } from 'graphql';
 
 const defaultOnRedirectCallback = (appState?: any): void => {
   window.location.replace(appState?.returnTo || window.location.pathname);
@@ -111,12 +111,12 @@ export class PhotonClientStore {
       state: {
         isLoading: boolean;
         data?: Prescription;
-        errors: GraphQLError[];
+        errors: GraphQLFormattedError[];
         error?: any;
       };
       createPrescription: (args: MutationCreatePrescriptionArgs) => Promise<{
         data: { createPrescription: Prescription } | null | undefined;
-        errors: readonly GraphQLError[] | undefined;
+        errors: readonly GraphQLFormattedError[] | undefined;
       }>;
     };
   };
@@ -159,7 +159,7 @@ export class PhotonClientStore {
       };
       prescription: {
         isLoading: boolean;
-        errors: GraphQLError[];
+        errors: GraphQLFormattedError[];
         error?: any;
         data?: Prescription;
       };
