@@ -201,7 +201,7 @@ export const PhotonMedicationDropdown = <T extends { id: string }>(props: {
   const rowVirtualizer = createMemo(() =>
     createVirtualizer({
       count: allItems().length,
-      getScrollElement: () => listRef,
+      getScrollElement: () => listRef ?? null,
       estimateSize: () => 36.8,
       overscan: 100
     })
@@ -435,7 +435,7 @@ const GroupLabelEl = (props: { item: GroupTitle }) => (
 );
 
 const ItemEl = (props: {
-  item: VirtualItem<unknown>;
+  item: VirtualItem;
   index: number;
   isLoader: boolean;
   isSelected: boolean;
