@@ -1,5 +1,5 @@
 import { PhotonClient } from '@photonhealth/sdk';
-import { GraphQLError } from 'graphql';
+import { GraphQLFormattedError } from 'graphql';
 import { createStore } from 'solid-js/store';
 import { DispenseUnit } from '@photonhealth/sdk/dist/types';
 
@@ -9,7 +9,7 @@ const createDispenseUnitStore = () => {
   const [store, setStore] = createStore<{
     dispenseUnits: {
       data: StoreDispenseUnit[];
-      errors: readonly GraphQLError[];
+      errors: readonly GraphQLFormattedError[];
       isLoading: boolean;
     };
   }>({
