@@ -19,7 +19,8 @@ import {
   REDROCK_STGEORGE_PHARMACY_ID,
   COST_PLUS_PHARMACY_ID,
   WALMART_MAIL_ORDER_PHARMACY_ID,
-  HEALTHWAREHOUSE_PHARMACY_ID
+  HEALTHWAREHOUSE_PHARMACY_ID,
+  STRIVE_PHARMACY_ID
 } from '../pharmacies';
 
 /**
@@ -86,7 +87,8 @@ const organizationSettings: {
       FOOTHILLS_PHARMACY_ID,
       SMARTSCRIPTS_PHARMACY_ID,
       OLYMPIA_PHARMACY_ID,
-      HEALTHWAREHOUSE_PHARMACY_ID
+      HEALTHWAREHOUSE_PHARMACY_ID,
+      STRIVE_PHARMACY_ID
     ],
     topRankedWalgreens: true,
     mailOrderNavigate: true,
